% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_project_docs.R
\name{mod_project_docs_server}
\alias{mod_project_docs_server}
\title{mod_project_docs_server}
\usage{
mod_project_docs_server(id, base_dir = "app_data/tarmac dashboard documents")
}
\arguments{
\item{id}{Shiny module ID}

\item{base_dir}{Character path to the root docs folder (default:
"app_data/tarmac dashboard documents")}
}
\description{
Server logic to list and download local PDFs.
}
