% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ai_forecast_fct_display.R
\name{predict_counts}
\alias{predict_counts}
\title{Predict expected arrivals for a site / CTAS / time window}
\usage{
predict_counts(
  model,
  hourly_df,
  site,
  ctas_vec = c(2, 3, 4),
  start_date,
  end_date,
  from_hour = 16,
  to_hour = 18,
  xlev = NULL
)
}
\description{
Predict expected arrivals for a site / CTAS / time window
}
